fix: resolve \$ref pointers in operation parameters
OpenAPI 3.x specs commonly declare parameters once in
components.parameters and reference them via \$ref from operations
(Stripe, Slack, GitHub, Sympla, etc all do this). The extractor was
only resolving \$ref for schemas, treating ref-only parameter entries
as raw objects with undefined name/in fields. This produced
'Cannot add option --undefined <undefined>' crashes.

Added resolveParameter() in a new refs.ts module and a guard against
parameters missing name/in after resolution. As part of the change,
split extractor.ts into focused modules (refs, params, naming) to
keep each file under 150 lines.
